ml_price_predictor/requirements.txt
# Core ML dependencies
catboost>=1.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ONNX support
onnx>=1.14.0
onnxruntime>=1.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
kaleido>=0.2.1  # For static image export

# Utilities
PyYAML>=6.0.0
tqdm>=4.65.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
pytest-cov>=4.1.0
kaleido>=0.2.1
